Claim task from the task list
